@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.4",
-    "@angular/cdk": "^6.2.1",
-    "@angular/common": "6.0.4",
-    "@angular/compiler": "6.0.4",
-    "@angular/core": "6.0.4",
-    "@angular/forms": "6.0.4",
-    "@angular/http": "6.0.4",
-    "@angular/material": "^6.2.1",
-    "@angular/platform-browser": "6.0.4",
-    "@angular/platform-browser-dynamic": "6.0.4",
-    "@angular/router": "6.0.4",
-    "@compodoc/compodoc": "^1.1.3",
+    "@angular/animations": "^6.1.4",
+    "@angular/cdk": "^6.1.4",
+    "@angular/common": "6.1.4",
+    "@angular/compiler": "6.1.4",
+    "@angular/core": "6.1.4",
+    "@angular/forms": "6.1.4",
+    "@angular/http": "6.1.4",
+    "@angular/material": "^6.1.4",
+    "@angular/platform-browser": "6.1.4",
+    "@angular/platform-browser-dynamic": "6.1.4",
+    "@angular/router": "6.1.4",
+    "@compodoc/compodoc": "^1.1.5",
     "bootstrap": "^3.3.7",
     "chart.js": "^2.7.1",
     "core-js": "^2.4.1",
@@ -38,8 +38,8 @@
     "ng2-charts": "^1.6.0",
     "ng2-eonasdan-datetimepicker": "^0.1.3",
     "ng2-order-pipe": "^0.1.5",
-    "rxjs": "^6.2.1",
-    "rxjs-compat": "^6.2.1",
+    "rxjs": "^6.1.4",
+    "rxjs-compat": "^6.1.4",
     "webpack": "^4.16.5",
     "zone.js": "^0.8.26"
   },
@@ -47,13 +47,12 @@
     "@angular-devkit/build-angular": "~0.6.8",
     "@angular-devkit/core": "0.6.8",
     "@angular/cli": "6.0.8",
-    "@angular/compiler-cli": "6.0.4",
-    "@angular/language-service": "6.0.4",
+    "@angular/compiler-cli": "6.1.4",
+    "@angular/language-service": "6.1.4",
     "@types/jasmine": "~2.8.4",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~9.3.0",
     "codelyzer": "^4.0.1",
-    "compodoc": "0.0.41",
     "jasmine-core": "~2.8.0",
     "jasmine-reporters": "^2.3.1",
     "jasmine-spec-reporter": "~4.2.1",
