@@ -3,10 +3,16 @@ import { browser, by, element } from 'protractor';
 export class Navbar {
 
     navigateTo(){
-        return browser.get('');
+        return browser.get('./#/login');
     }
 
-    goToHome(){
+    logIn(username, password){
+      element(by.name('username')).sendKeys(username);
+      element(by.name('password')).sendKeys(password);
+      element(by.buttonText('Sign in')).click();
+    }
+
+    goToAdminHome(){
       element(by.linkText('Home')).click();
         //element(by.css('body > app-component > app-navbar > nav > div > ul:nth-child(2) > li:nth-child(1)')).click();
 
@@ -41,7 +47,18 @@ export class Navbar {
       element(by.linkText('Predictions')).click();
         //element(by.css('body > app-component > app-navbar > nav > div > ul:nth-child(2) > li:nth-child(5)')).click();
     }
+
+    goToSalesForce(){
+      element(by.linkText('Salesforce')).click();
+    }
     getCurrentURL(){
         return browser.getCurrentUrl();
+    }
+
+    goToAssociateHome(){
+      element(by.linkText('Home')).click();
+    }
+    goToMyInterview(){
+      element(by.xpath('/html/body/app-component/app-navbar/nav/div/ul[2]/li[2]/a')).click();
     }
 }
