@@ -7,13 +7,13 @@ const { SpecReporter } = require('jasmine-spec-reporter');
 exports.config = {
   allScriptsTimeout: 20000,
   specs: [
-     './e2e/register-associate/register-associate.e2e-spec.ts',
-      './e2e/login/login.e2e-spec.ts',
+    './e2e/register-associate/register-associate.e2e-spec.ts',
+    './e2e/login/login.e2e-spec.ts',
 
-     './e2e/navbar/navbar.e2e-spec.ts',
-     './e2e/create-user/create-user.e2e-spec.ts',
-    './e2e/predictions/predictions.e2e-spec.ts',
-    //  './e2e/client-list/client-list-spec.ts',
+    './e2e/navbar/navbar.e2e-spec.ts',
+    './e2e/create-user/create-user.e2e-spec.ts',
+    //'./e2e/predictions/predictions.e2e-spec.ts',
+    './e2e/client-list/client-list-spec.ts',
     //   './e2e/app/app.e2e-spec.ts',
     //   './e2e/associate-list/associate-list.js',
     //  './e2e/associate-list/associate-list-search.js',
@@ -31,8 +31,8 @@ exports.config = {
     }
   },
   // directConnect: true,
-    //baseUrl: 'http://localhost:4200/',
-   baseUrl: 'http://34.227.178.103:8090/NGTrackForce/',
+   //baseUrl: 'http://localhost:4200/',
+   baseUrl: 'http://34.227.178.103:8090/NGTrackForce',
    //for pipeline delpoyment
      seleniumAddress: "http://127.0.0.1:4444/wd/hub",
     // SELENIUM_PROMISE_MANAGER: false,
