@@ -28,7 +28,7 @@ const BATCHES_COTRAINER_KEY = 'currentBatchesCotrainer';
 const BATCHES_KEY = 'currentBatches';
 const CLIENTS_KEY = 'currentClients';
 
-/* 
+/*
   PROBLEM
   should be admin/sales/staging_key
   SHOULDNT load all associates as soon as one of those roles logs in
@@ -79,9 +79,9 @@ export class LoginComponent implements OnInit {
   public newTrainer: Trainer;
   public newAssociate: Associate;
 
-  public isLoggingIn: boolean = false;
-  public loginClicked: boolean = false;
-
+  public isLoggingIn = false;
+  public loginClicked = false;
+  public userRequirement: String = ""
   /**
    *@constructor
    *
@@ -127,7 +127,7 @@ export class LoginComponent implements OnInit {
     if (user != null) {
       this.loginClicked = true;
       this.isLoggingIn = true;
-      
+
 
       this.userService.checkJwtValid().subscribe(
         data => { this.routeToUserHome(user.role); },
@@ -139,11 +139,11 @@ export class LoginComponent implements OnInit {
   routeToUserHome(role: number){
     this.navbarService.show();
 
-    if (role == 5) {
+    if (role === 5) {
         this.router.navigate(['associate-view']);
-    } else if (role == 2) {
+    } else if (role === 2) {
         this.router.navigate(['trainer-view']);
-    } else if (role == 1 || role == 3 || role == 4) {
+    } else if (role === 1 || role === 3 || role === 4) {
         this.router.navigate(['app-home']);
     } else{
       this.navbarService.hide();
@@ -175,13 +175,14 @@ export class LoginComponent implements OnInit {
     this.sucMsg = "";
     this.errMsg = "";
 
+
     if (this.password === undefined || this.cpassword === undefined || this.password.length === 0 || this.cpassword.length === 0) {
       this.errMsg = 'Please enter a password and confirm password!';
     } else if (this.password !== this.cpassword) {
       this.errMsg = 'Passwords do not match!';
     } else if (!this.usernameRestrictions.test(this.username.trim()) && !this.passwordRestrictions.test(this.password)) {
       this.errMsg = `Invalid username and password!<br>
-        <br>Password requirements:
+        <br>Username requirements:
           <ul>
             <li>• 6 - 20 alphanumeric characters.</li>
             <li>• No spaces.</li>
@@ -197,7 +198,7 @@ export class LoginComponent implements OnInit {
           </ul>`;
     } else if (!this.usernameRestrictions.test(this.username.trim())) {
       this.errMsg = `Invalid username!<br>
-        <br>Password requirements:
+        <br>Username requirements:
           <ul>
             <li>• 6 - 20 characters.</li>
             <li>• No spaces.</li>
