@@ -1,9 +1,9 @@
 /**
- * Decorator for automatically unsubscribing from any and all observables (so that you don't need to bloat 
+ * Decorator for automatically unsubscribing from any and all observables (so that you don't need to bloat
  *  your codebase by having instances of said observables floating around solely for the purpose of manually
  *  unsubscribing!)
- * @param constructor 
- * 
+ * @param constructor
+ *
  * - This automagical code obtained from https://netbasal.com/automagically-unsubscribe-in-angular-4487e9853a88
  * - There is also a github repository https://github.com/NetanelBasal/ngx-auto-unsubscribe
  * - It's possible to just npm install a dependency and get the same functionality
@@ -15,11 +15,11 @@ export function AutoUnsubscribe( constructor ) {
 
     // make copy of the ngOnDestroy callback
     const original = constructor.prototype.ngOnDestroy;
-  
+
     // upon destruction of the component...
     constructor.prototype.ngOnDestroy = function () {
       // ...iterate thru its properties...
-      for ( let prop of this ) {
+      for ( const prop of this ) {
         const property = this[ prop ];
         // if property exists and has-a unsubscribe...
         if ( property && (typeof property.unsubscribe === "function") ) {
@@ -29,7 +29,8 @@ export function AutoUnsubscribe( constructor ) {
       }
       // if the callback we copied was function, we invoke it now with any arguments that were passed in
       //  via constructor
+      // tslint:disable-next-line:no-unused-expression
       original && typeof original === "function" && original.apply(this, arguments);
     };
-   
+
 }
