@@ -6,15 +6,15 @@
 
 export class StatusInfo {
   name: string;
-  trainingMapped: number;
-  trainingUnmapped: number;
-  reservedMapped: number;
-  openUnmapped: number;
-  selectedMapped: number;
-  selectedUnmapped: number;
-  confirmedMapped: number;
-  confirmedUnmapped: number;
-  deployedMapped: number;
-  deployedUnmapped: number;
+  trainingMapped: number; //1
+  trainingUnmapped: number; //6
+  reservedMapped: number; //2
+  openUnmapped: number; //7
+  selectedMapped: number; //3
+  selectedUnmapped: number; //8
+  confirmedMapped: number; //4
+  confirmedUnmapped: number; //9
+  deployedMapped: number; //5
+  deployedUnmapped: number; //10
   constructor() { }
 }
