@@ -5,7 +5,7 @@ import { Pipe, PipeTransform } from '@angular/core';
 })
 
 /**
- * Pipe to filter searches based on select option of Client 
+ * Pipe to filter searches based on select option of Client
  * @author Cyril Mathew
  */
 export class AssociateSearchByClientPipe implements PipeTransform {
@@ -13,15 +13,15 @@ export class AssociateSearchByClientPipe implements PipeTransform {
 
         if (!items) {
             return [];
-        } 
-        
+        }
+
         if (!searchText) {
             return items;
-        } 
-        
+        }
+
         //return results that contain firstname, lastname, status, client, id that matches a client selection
         return items.filter(associate => {
-            return ((associate.client != null)? associate.client.name == searchText: false);
+            return ((associate.client != null)? associate.client.name === searchText: false);
         });
     }
 }
