@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, AsyncSubject, BehaviorSubject } from 'rxjs';
 
 import { Associate } from '../../models/associate.model';
 import { environment } from '../../../environments/environment';
@@ -15,19 +15,79 @@ import { GraphCounts } from '../../models/graph-counts';
 export class AssociateService {
   private baseURL: string = environment.url + 'TrackForce/associates';
   private nassURL: string = this.baseURL + '/nass';
+
+  /**
+   * These behavior subjects will hold an initial empty value until a request is sent for their
+   * data. This should not be the way it is implemented however since we have to remain compatible 
+   * with the components written already sending a null value from an AsyncSubject would throw many errors
+   *
+   * Each of the private variables below represents the return of each of the functions that are in this
+   * service this needs to be refactored but in this first pass of the service revamp top priority is to
+   * stop having the data reload each time and allow the application to hold a copy of the data to load
+   * then update in the background.
+   * 
+   * NOTE: The $ on a variable mean it is an Observable
+   * 
+   * TODO: switch to using an AsyncSubject
+   *
+   * Note on things to do that would then allow this
+   * 1. In the dependant components null checking will be needed
+   * 2. Once components can handle a null value switch the BehaviorSubject to an Async Subject
+   * 3. Ideally the components will show loading, error, and success by listening to the three
+   *    events that a subject will output when subscribed to
+   */
+  private allAssociates$: BehaviorSubject<Associate[]> = new BehaviorSubject<Associate[]>([]);
+
+  private associateCount$: BehaviorSubject<number[]> = new BehaviorSubject<number[]>([]);
+
+  private updateAssociates$: AsyncSubject<boolean> = new AsyncSubject<boolean>();
+
+  private updateAssociate$: AsyncSubject<boolean> = new AsyncSubject<boolean>();
+
+  private approveAssociate$: AsyncSubject<boolean> = new AsyncSubject<boolean>();
+
+  private getAssociatesByStatus$: AsyncSubject<GraphCounts[]> = new AsyncSubject<GraphCounts[]>();
+
+  private getUndeployedAssociates$: AsyncSubject<GraphCounts[]> = new AsyncSubject<GraphCounts[]>();
+
+  // TODO: Decide if empty strings is better or if a loading message should be put here
+  // TODO: Why is there a get by user and get by associate this is two different models on backend
+  // ? should one be handled by another service
+  // ? Do we really need two in this section of the frontend
+  private associateByUserId$: BehaviorSubject<Associate> = new BehaviorSubject<Associate>(new Associate(
+    "",
+    "",
+    null
+  ));
+
+  private associateByAssociateId$: BehaviorSubject<Associate> = new BehaviorSubject<Associate>(new Associate(
+    "",
+    "",
+    null
+  ));
+
   constructor(private http: HttpClient) {}
 
   /**
    *
    * Gets all of the associates
    */
-  getAllAssociates(): Observable<Associate[]> {
+  getAllAssociates(): BehaviorSubject<Associate[]> {
     const url: string = this.baseURL + '/allAssociates';
-    return this.http.get<Associate[]>(url);
+    this.http.get<Associate[]>(url).subscribe(
+      (data: Associate[]) => this.allAssociates$.next(data),
+      error => this.allAssociates$.error(error)
+    );
+    return this.allAssociates$;
   }
 
-  /*
+  /**
     gets initial associates loaded
+
+    ! DEPRECATED: This is not a good idea to try to speed up performance
+    ! the plan will be to implement pagination for now the app will just
+    ! wait for all associates as this request will end up taking just as
+    ! long due to the problems on the server side
   */
   getNAssociates(): Observable<Associate[]> {
     return this.http.get<Associate[]>(this.nassURL);
@@ -35,9 +95,13 @@ export class AssociateService {
   /**
    * get the count of the associates to display in the pie charts on the home page
    */
-  getCountAssociates(): Observable<number[]> {
+  getCountAssociates(): BehaviorSubject<number[]> {
     const url: string = this.baseURL + '/countAssociates';
-    return this.http.get<number[]>(url);
+    this.http.get<number[]>(url).subscribe(
+      (data: number[]) => this.associateCount$.next(data),
+      error => this.associateCount$.error(error)
+    );
+    return this.associateCount$;
   }
 
   /**
@@ -47,7 +111,11 @@ export class AssociateService {
    */
   getAssociate(id: number) {
     const url: string = this.baseURL + '/' + id;
-    return this.http.get<Associate>(url);
+    this.http.get<Associate>(url).subscribe(
+      (data: Associate) => this.associateByUserId$.next(data),
+      error => this.associateByUserId$.error(error)
+    );
+    return this.associateByUserId$;
   }
 
   /**
@@ -57,7 +125,11 @@ export class AssociateService {
    */
   getByAssociateId(id: number) {
     const url: string = this.baseURL + '/associates/' + id;
-    return this.http.get<Associate>(url);
+    this.http.get<Associate>(url).subscribe(
+      (data: Associate) => this.associateByAssociateId$.next(data),
+      error => this.associateByAssociateId$.error(error)
+    );
+    return this.associateByAssociateId$;
   }
 
   /**
@@ -66,12 +138,24 @@ export class AssociateService {
    * @param ids - list of associate ids of associates to be updated
    * @param marketingStatusId - the marketing status these associates will be updated to
    * @param clientId - the client id that the associates will be mapped to
+   * 
+   * ? Changing this to use BehaviorSubjects however it may be wanted to have it not multiplex
    */
   updateAssociates(ids: number[], verification: number, marketingStatusId: number, clientId: number): Observable<boolean> {
     const url: string = this.baseURL + '?marketingStatusId=' + marketingStatusId + '&clientId=' + clientId + '&verification=' + verification;
-    return this.http.put<boolean>(url, ids);
+    this.http.put<boolean>(url, ids).subscribe(
+      (data) => {
+        this.updateAssociates$.next(data);
+      },
+      (error) => {
+        this.updateAssociates$.error(error);
+      }
+    );
+
+    return  this.updateAssociates$;
   }
 
+  
   /**
    *
    * This method updates the associate in the database
@@ -79,20 +163,40 @@ export class AssociateService {
    */
   updateAssociate(associate: any) {
     const url: string = this.baseURL + '/' + associate.id;
-    return this.http.put<boolean>(url, associate);
+    this.http.put<boolean>(url, associate).subscribe(
+      data => this.updateAssociate$.next(data),
+      error => this.updateAssociate$.error(error)
+    );
+    return this.updateAssociate$;
   }
 
+  // ? Below this point the requests are for stats which may need to become a seperate service
+  // For now that decision can be left to the service revamp for now the service update will
+  // focus on keeping one copy of data to aid performance
+
   getAssociatesByStatus(statusId: number): Observable<GraphCounts[]> {
-    return this.http.get<GraphCounts[]>(this.baseURL + '/mapped/' + statusId);
+    this.http.get<GraphCounts[]>(this.baseURL + '/mapped/' + statusId).subscribe(
+      (data: GraphCounts[]) => this.getAssociatesByStatus$.next(data),
+      error => this.getAssociatesByStatus$.error(error)
+    );
+    return this.getAssociatesByStatus$;
   }
 
   approveAssociate(associateID: number) {
     const url: string = this.baseURL + '/' + associateID + '/approve';
-    return this.http.put<boolean>(url, associateID);
+    this.http.put<boolean>(url, associateID).subscribe(
+      data => this.approveAssociate$.next(data),
+      error => this.approveAssociate$.error(error)
+    );
+    return  this.approveAssociate$;
   }
 
   getUndeployedAssociates(mappedOrUnmapped: string): Observable<GraphCounts[]> {
     const url: string = this.baseURL + '/undeployed/' + mappedOrUnmapped;
-    return this.http.get<GraphCounts[]>(url);
+    this.http.get<GraphCounts[]>(url).subscribe(
+      (data: GraphCounts[]) => this.getUndeployedAssociates$.next(data),
+      error => this.getUndeployedAssociates$.error(error)
+    );
+    return this.getUndeployedAssociates$;
   }
 }
