@@ -1,13 +1,13 @@
 package com.revature.dao;
-
 import java.util.List;
-
 import com.revature.criteria.GraphedCriteriaResult;
 import com.revature.entity.TfAssociate;
 
 public interface AssociateDao {
 
 	TfAssociate getAssociate(Integer associateid);
+	
+	List<TfAssociate> getNAssociateMatchingCriteria(int startIdx, int numRes, int mktStatus, int clientId);
 
 	List<TfAssociate> getAllAssociates();
 	
