@@ -1,23 +1,16 @@
 package com.revature.daoimpl;
-
 import static com.revature.utils.HibernateUtil.runHibernateTransaction;
 import static com.revature.utils.HibernateUtil.saveToDB;
-
-import java.math.BigDecimal;
 import java.util.List;
-
 import javax.persistence.Query;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Join;
 import javax.persistence.criteria.Path;
 import javax.persistence.criteria.Root;
-import javax.persistence.Entity;
-
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.openqa.selenium.InvalidArgumentException;
-
 import com.revature.criteria.GraphedCriteriaResult;
 import com.revature.dao.AssociateDao;
 import com.revature.entity.TfAssociate;
@@ -28,27 +21,63 @@ import com.revature.entity.TfMarketingStatus;
 import com.revature.entity.TfUser;
 import com.revature.utils.HibernateUtil;
 import com.revature.utils.Sessional;
+
 /**
- * Data Access Object implementation to access the associate entity from the
- * Database
+ * Data Access Object implementation to access the associate entity from the Database
  */
 public class AssociateDaoImpl implements AssociateDao {
 	
+	/** Gets list of associates matching criteria. Used by updated angular front end to perform 
+	 * Pagination of results and improve performance.
+	 * @author Joshua-Pressley-1807
+	 * @param startIdx starting index
+	 * @param numRes the number of resuts to return
+	 * @param mktStatus the marketing ID
+	 * @param clientId the client ID
+	 * @return list of associates matching criteria */
+	public List<TfAssociate> getNAssociateMatchingCriteria(int startIdx, int numRes, int mktStatus, int clientId)
+	{
+		Session session = HibernateUtil.getSessionFactory().openSession();
+		CriteriaBuilder builder = session.getCriteriaBuilder();
+		CriteriaQuery<TfAssociate> criteria = builder.createQuery(TfAssociate.class);
+		Root<TfAssociate> root = criteria.from(TfAssociate.class);
+		List<TfAssociate> results;
+		
+		if (clientId == -1 && mktStatus != -1) {
+			criteria.where(builder.equal(root.get("marketingStatus"), mktStatus));
+		} 
+		else if (mktStatus == -1 && clientId != -1) {
+			criteria.where(builder.equal(root.get("client"), clientId));
+		} 
+		else if (mktStatus != -1 && clientId != -1) {
+			criteria.where(builder.and(
+					builder.equal(root.get("marketingStatus"), mktStatus),
+					builder.equal(root.get("client"), clientId)));
+		}
+		
+		if (startIdx==1) { startIdx = 0; }		
+		results = session.createQuery(criteria)
+				.setFirstResult(startIdx)
+				.setMaxResults(numRes)
+				.getResultList();
+		
+		session.close(); //close out the session
+		if (results == null || results.size() == 0) { return null; }
+		return results;
+	}//end getNAssociateMatchingCriteria()
+	
 	/**
 	 * Gets a single associate with an id
-	 * 
 	 * @param Integer associateId
 	 */
 	@Override
 	public TfAssociate getAssociate(Integer id) {
 		return HibernateUtil.runHibernate((Session session, Object ... args) ->
 		session.createQuery("from TfAssociate a where a.id = :id", TfAssociate.class).setParameter("id", id).getSingleResult());
-
 	}
 
 	/**
 	 * Gets an associate by an associated user id
-	 * 
 	 * @param int userId
 	 */
 	@Override
@@ -75,282 +104,120 @@ public class AssociateDaoImpl implements AssociateDao {
 				.getResultList());
 	}
 	
-	@Override
-	public Object getCountUndeployedMapped()
-	{
+	/** getCount uses CriteriaQuery and cascading switch statement to return count which 
+	 *  matche the marketingStatus. Special query for cases 11 and 12
+	 *  @author Paul Capellan - 1807
+	 *  @param tfmsid is marketing status to be compared
+	 *  @return returns count based on marketingStatus equivalent to tfmsid for case 1 to 10,
+	 *  or statement used for case 11 and 12
+	 */
+	//Called by all getCount methods
+	public Object getCount(int tfmsid) {
 		Session session = null;
-		Object undeployedmapped = null;
-		
+		Object count = null;
+		 
 		try {
 			session = HibernateUtil.getSessionFactory().openSession();
-			undeployedmapped = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where (tf_marketing_status_id = 1 or tf_marketing_status_id = 2 or tf_marketing_status_id = 3 " +
-					    "or tf_marketing_status_id = 4)"
-					).getSingleResult();
+			
+			CriteriaBuilder builder = session.getCriteriaBuilder();
+			CriteriaQuery<Long> qr = builder.createQuery(Long.class);
+			Root<TfAssociate> root = qr.from(TfAssociate.class);
+			List<Long> results;
+			qr.select(builder.count(root));
+			
+			switch(tfmsid) {	
+				case 11: 
+						qr.where(builder.or(builder.or(builder.equal(root.get("marketingStatus"), 1), 
+								builder.equal(root.get("marketingStatus"), 2)), builder.or(builder.equal(
+								root.get("marketingStatus"), 3), builder.equal(root.get("marketingStatus"), 4))));
+						break;
+				case 12:  
+						qr.where(builder.or(builder.or(builder.equal(root.get("marketingStatus"), 6), 
+								builder.equal(root.get("marketingStatus"), 7)), builder.or(builder.equal(
+								root.get("marketingStatus"), 8), builder.equal(root.get("marketingStatus"), 9))));
+						break;
+				default:
+						qr.where(builder.equal(root.get("marketingStatus"), tfmsid));
+						break;
+			}//end switch
+			
+			results = session.createQuery(qr).getResultList();
+			count = results.get(0);
+			//System.out.println(tfmsid + " > " + count);
 		}catch(HibernateException e) {
 			e.printStackTrace();
 		}finally {
 			if ( session != null )
-			{
 				session.close();
-			}
 		}
-		return undeployedmapped;
+		return count;
+	}
+	
+	/** Calls the getCount method above and passes respective parameter
+	 *  First 2 methods preserve the same query of tf_marketing_status_id, 11 or 12 not valid marketingStatus 
+	 *  @author Paul Capellan - 1807
+	 *  @return returns count from getCount() method based on marketingStatus
+	 */
+	//Call by getCount(11), but preserve the same tf_marketing_status_id!
+	@Override
+	public Object getCountUndeployedMapped() {
+		return getCount(11);
+	}
+	
+	//Call by getCount(12), but preserve the same tf_marketing_status_id!
+	@Override
+	public Object getCountUndeployedUnmapped() {
+		return getCount(12);
 	}
 	
 	@Override
-	public Object getCountUndeployedUnmapped()
-	{
-		Session session = null;
-		Object undeployedunmapped = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			undeployedunmapped = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where (tf_marketing_status_id = 6 or tf_marketing_status_id = 7 " +
-					    "or tf_marketing_status_id = 8 or tf_marketing_status_id = 9)"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return undeployedunmapped;
+	public Object getCountDeployedMapped() {
+		return getCount(5);
 	}
 	
 	@Override
-	public Object getCountDeployedMapped()
-	{
-		Session session = null;
-		Object deployedmapped = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			deployedmapped = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 5"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return deployedmapped;
-	}
-	
-	@Override
-	public Object getCountDeployedUnmapped()
-	{
-		Session session = null;
-		Object deployedunmapped = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			deployedunmapped = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 10"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return deployedunmapped;
+	public Object getCountDeployedUnmapped() {
+		return getCount(10);
 	}
 
 	@Override
-	public Object getCountUnmappedTraining()
-	{
-		Session session = null;
-		Object unmappedtraining = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			unmappedtraining = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 6"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return unmappedtraining;
+	public Object getCountUnmappedTraining() {
+		return getCount(6);
 	}
 	
 	@Override
-	public Object getCountUnmappedOpen()
-	{
-		Session session = null;
-		Object unmappedopen = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			unmappedopen = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 7"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return unmappedopen;
+	public Object getCountUnmappedOpen() {
+		return getCount(7);
 	}
 	
 	@Override
-	public Object getCountUnmappedSelected()
-	{
-		Session session = null;
-		Object unmappedselected = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			unmappedselected = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 8"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return unmappedselected;
+	public Object getCountUnmappedSelected() {
+		return getCount(8);
 	}
 	
 	@Override
-	public Object getCountUnmappedConfirmed()
-	{
-		Session session = null;
-		Object unmappedconfirmed = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			unmappedconfirmed = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 9"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return unmappedconfirmed;
+	public Object getCountUnmappedConfirmed() {
+		return getCount(9);
 	}
 	
 	@Override
-	public Object getCountMappedTraining()
-	{
-		Session session = null;
-		Object mappedtraining = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			mappedtraining = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 1"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return mappedtraining;
+	public Object getCountMappedTraining() {
+		return getCount(1);
 	}
 	
 	@Override
-	public Object getCountMappedReserved()
-	{
-		Session session = null;
-		Object mappedreserved = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			mappedreserved = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 2"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return mappedreserved;
+	public Object getCountMappedReserved() {
+		return getCount(2);
 	}
 	
 	@Override
-	public Object getCountMappedSelected()
-	{
-		Session session = null;
-		Object mappedselected = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			mappedselected = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 3"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return mappedselected;
+	public Object getCountMappedSelected() {
+		return getCount(3);
 	}
 	
 	@Override
-	public Object getCountMappedConfirmed()
-	{
-		Session session = null;
-		Object mappedconfirmed = null;
-		
-		try {
-			session = HibernateUtil.getSessionFactory().openSession();
-			mappedconfirmed = session.createNativeQuery(
-					"select count(tf_associate_id) from admin.tf_associate " +
-					"where tf_marketing_status_id = 4"
-					).getSingleResult();
-		}catch(HibernateException e) {
-			e.printStackTrace();
-		}finally {
-			if ( session != null )
-			{
-				session.close();
-			}
-		}
-		return mappedconfirmed;
+	public Object getCountMappedConfirmed() {
+		return getCount(4);
 	}
 
 	@Override
@@ -429,54 +296,48 @@ public class AssociateDaoImpl implements AssociateDao {
 			return session.createQuery(query).getResultList();
 		}, id);
 	}
-
+	
+	/** Optimized getUndeployed to remove redundancy
+	 * @author Paul C.-1807*/
 	@Override
 	public List<GraphedCriteriaResult> getUndeployed(String which) {
-		if (which.equals("mapped")) {
+		if(which.equals("mapped") || which.equals("unmapped")) {
 			return HibernateUtil.runHibernate((Session session, Object... args) -> {
 				CriteriaBuilder cb = session.getCriteriaBuilder();
 				CriteriaQuery<GraphedCriteriaResult> query = cb.createQuery(GraphedCriteriaResult.class);
-
 				Root<TfAssociate> root = query.from(TfAssociate.class);
-
-				Join<TfAssociate, TfClient> clientJoin = root.join("client");
-				Join<TfAssociate, TfMarketingStatus> msJoin = root.join("marketingStatus");
-
-				Path<?> clientId = clientJoin.get("id");
-				Path<?> clientName = clientJoin.get("name");
-
-				query.where(cb.lessThanOrEqualTo(msJoin.get("id"), 4));
-				query.where(cb.greaterThanOrEqualTo(msJoin.get("id"), 1));
-
-				query.groupBy(clientId, clientName);
-				query.multiselect(cb.count(root), clientId, clientName);
-				return session.createQuery(query).getResultList();
-			});
-		} else if (which.equals("unmapped")) {
-			return HibernateUtil.runHibernate((Session session, Object... args) -> {
-				CriteriaBuilder cb = session.getCriteriaBuilder();
-				CriteriaQuery<GraphedCriteriaResult> query = cb.createQuery(GraphedCriteriaResult.class);
-
-				Root<TfAssociate> root = query.from(TfAssociate.class);
-
-				Join<TfAssociate, TfBatch> batchJoin = root.join("batch");
-				Join<TfBatch, TfCurriculum> curriculumJoin = batchJoin.join("curriculumName");
-				Join<TfAssociate, TfMarketingStatus> msJoin = root.join("marketingStatus");
-
-				Path<?> curriculumid = curriculumJoin.get("id");
-				Path<?> curriculumName = curriculumJoin.get("name");
-
-				query.where(cb.lessThanOrEqualTo(msJoin.get("id"), 9));
-				query.where(cb.greaterThanOrEqualTo(msJoin.get("id"), 6));
-				query.groupBy(curriculumid, curriculumName);
-				query.multiselect(cb.count(root), curriculumid, curriculumName);
+				
+				if (which.equals("mapped")) {
+					Join<TfAssociate, TfClient> clientJoin = root.join("client");
+					Join<TfAssociate, TfMarketingStatus> msJoin = root.join("marketingStatus");
+		
+					Path<?> clientId = clientJoin.get("id");
+					Path<?> clientName = clientJoin.get("name");
+		
+					query.where(cb.lessThanOrEqualTo(msJoin.get("id"), 4));
+					query.where(cb.greaterThanOrEqualTo(msJoin.get("id"), 1));
+		
+					query.groupBy(clientId, clientName);
+					query.multiselect(cb.count(root), clientId, clientName);
+				}
+				else if (which.equals("unmapped")) {
+					Join<TfAssociate, TfBatch> batchJoin = root.join("batch");
+					Join<TfBatch, TfCurriculum> curriculumJoin = batchJoin.join("curriculumName");
+					Join<TfAssociate, TfMarketingStatus> msJoin = root.join("marketingStatus");
+		
+					Path<?> curriculumid = curriculumJoin.get("id");
+					Path<?> curriculumName = curriculumJoin.get("name");
+		
+					query.where(cb.lessThanOrEqualTo(msJoin.get("id"), 9));
+					query.where(cb.greaterThanOrEqualTo(msJoin.get("id"), 6));
+					query.groupBy(curriculumid, curriculumName);
+					query.multiselect(cb.count(root), curriculumid, curriculumName);
+				}
 				return session.createQuery(query).getResultList();
 			});
 		}
-		
-		throw new InvalidArgumentException("NOT MAPPED OR UNMAPPED YOU FOOOL");
-	}
-
+		throw new InvalidArgumentException("getUndeployed(): NOT MAPPED OR UNMAPPED");
+	}//end getUndeployed()
 
 	@Override
 	public boolean updateAssociate(TfAssociate associate) {
