@@ -1,5 +1,4 @@
 package com.revature.entity;
-// Generated Nov 7, 2017 9:24:46 PM by Hibernate Tools 5.2.5.Final
 
 import java.sql.Timestamp;
 import java.util.HashSet;
@@ -9,11 +8,14 @@ import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
+import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
@@ -43,17 +45,16 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 @XmlRootElement
 @Entity
 @Table(name = "TF_ASSOCIATE", schema = "ADMIN")
-//@Cache(usage=CacheConcurrencyStrategy.READ_WRITE)
-// Logans attempt at getting ehcache working below
 @Cache(usage=CacheConcurrencyStrategy.READ_WRITE, region="TrackForce")
 public class TfAssociate implements java.io.Serializable {
-
 
 	private static final long serialVersionUID = -2324082555924677252L;
 
 	@XmlElement
 	@Id
 	@Column(name = "TF_ASSOCIATE_ID", unique = true, nullable = false, precision = 22, scale = 0)
+	@SequenceGenerator(name="associdSeq", sequenceName="associd_seq")
+	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="associdSeq")
 	private Integer id;
 
 	@XmlElement
@@ -89,7 +90,6 @@ public class TfAssociate implements java.io.Serializable {
 	@Column(name = "TF_ASSOCIATE_LAST_NAME", length = 30)
 	private String lastName;
 
-//	@JsonIgnore
 	@XmlElement
 	@OneToMany(fetch = FetchType.LAZY, mappedBy = "associate", cascade = {CascadeType.ALL})
 	@JsonIgnore
@@ -108,9 +108,9 @@ public class TfAssociate implements java.io.Serializable {
 	@Column(name = "TF_STAGING_FEEDBACK")
 	private String stagingFeedback;
 	
-
-	public TfAssociate() {}
-
+	public TfAssociate() {
+		super();
+	}
 
 	public TfAssociate(Integer id, TfUser user, TfBatch batch, TfMarketingStatus marketingStatus, TfClient client,
 			TfEndClient endClient, String firstName, String lastName, Set<TfInterview> interview,
@@ -129,93 +129,65 @@ public class TfAssociate implements java.io.Serializable {
 		this.clientStartDate = clientStartDate;
 	}
 
-	
-
 	public String getStagingFeedback() {
 		return stagingFeedback;
 	}
-
-
 	public void setStagingFeedback(String stagingFeedback) {
 		this.stagingFeedback = stagingFeedback;
 	}
 
-
 	public Integer getId() {
 		return id;
 	}
-
-
 	public void setId(Integer id) {
 		this.id = id;
 	}
 
-
 	public TfUser getUser() {
 		return user;
 	}
-
-
 	public void setUser(TfUser user) {
 		this.user = user;
 	}
 
-
 	public TfBatch getBatch() {
 		return batch;
 	}
-
-
 	public void setBatch(TfBatch batch) {
 		this.batch = batch;
 	}
 
-
 	public TfMarketingStatus getMarketingStatus() {
 		return marketingStatus;
 	}
-
-
 	public void setMarketingStatus(TfMarketingStatus marketingStatus) {
 		this.marketingStatus = marketingStatus;
 	}
 
-
 	public TfClient getClient() {
 		return client;
 	}
-
-
 	public void setClient(TfClient client) {
 		this.client = client;
 	}
 
-
 	public TfEndClient getEndClient() {
 		return endClient;
 	}
-
-
 	public void setEndClient(TfEndClient endClient) {
 		this.endClient = endClient;
 	}
 
-
 	public String getFirstName() {
 		return firstName;
 	}
-
-
 	public void setFirstName(String firstName) {
 		this.firstName = firstName;
 	}
 
-
 	public String getLastName() {
 		return lastName;
 	}
-
-
 	public void setLastName(String lastName) {
 		this.lastName = lastName;
 	}
@@ -234,28 +206,21 @@ public class TfAssociate implements java.io.Serializable {
 	public Set<TfPlacement> getPlacement() {
 		return placement;
 	}
-
 	@JsonIgnore
 	public void setPlacement(Set<TfPlacement> placement) {
 		this.placement = placement;
 	}
 
-
 	public Timestamp getClientStartDate() {
 		return clientStartDate;
 	}
-
-
 	public void setClientStartDate(Timestamp clientStartDate) {
 		this.clientStartDate = clientStartDate;
 	}
 
-
 	public static long getSerialversionuid() {
 		return serialVersionUID;
 	}
-
-	
 
 	@Override
 	public String toString() {
@@ -263,8 +228,4 @@ public class TfAssociate implements java.io.Serializable {
 				+ client + ", endClient=" + endClient + ", firstName=" + firstName + ", lastName=" + lastName
 				+ ", clientStartDate=" + clientStartDate + ", stagingFeedback=" + stagingFeedback + "]";
 	}
-
-
-
-		
 }
