@@ -1,5 +1,4 @@
 package com.revature.entity;
-// Generated Nov 7, 2017 9:24:46 PM by Hibernate Tools 5.2.5.Final
 
 import java.util.HashSet;
 import java.util.Set;
@@ -21,8 +20,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 @XmlRootElement
 @Entity
 @Table(name = "TF_END_CLIENT", schema = "ADMIN")
-//@Cache(usage=CacheConcurrencyStrategy.READ_WRITE)
-//Logans attempt at getting ehcache working below
 @Cache(usage=CacheConcurrencyStrategy.READ_WRITE, region="TrackForce")
 public class TfEndClient implements java.io.Serializable {
 
@@ -30,7 +27,6 @@ public class TfEndClient implements java.io.Serializable {
 	
 	@XmlElement
 	@Id
-//	@GeneratedValue
 	@Column(name = "TF_END_CLIENT_ID", unique = true, nullable = false, precision = 22, scale = 0)
 	private Integer id;
 	
@@ -51,6 +47,7 @@ public class TfEndClient implements java.io.Serializable {
 	private Set<TfInterview> interviews = new HashSet<TfInterview>(0);
 
 	public TfEndClient() {
+		super();
 	}
 
 	public TfEndClient(Integer id, String name, Set<TfAssociate> associates, Set<TfPlacement> placements,
@@ -66,7 +63,6 @@ public class TfEndClient implements java.io.Serializable {
 	public Integer getId() {
 		return id;
 	}
-
 	public void setId(Integer id) {
 		this.id = id;
 	}
@@ -74,7 +70,6 @@ public class TfEndClient implements java.io.Serializable {
 	public String getName() {
 		return name;
 	}
-
 	public void setName(String name) {
 		this.name = name;
 	}
@@ -83,7 +78,6 @@ public class TfEndClient implements java.io.Serializable {
 	public Set<TfAssociate> getAssociates() {
 		return associates;
 	}
-
 	@JsonIgnore
 	public void setAssociates(Set<TfAssociate> associates) {
 		this.associates = associates;
@@ -93,7 +87,6 @@ public class TfEndClient implements java.io.Serializable {
 	public Set<TfPlacement> getPlacements() {
 		return placements;
 	}
-
 	@JsonIgnore
 	public void setPlacements(Set<TfPlacement> placements) {
 		this.placements = placements;
@@ -103,7 +96,6 @@ public class TfEndClient implements java.io.Serializable {
 	public Set<TfInterview> getInterviews() {
 		return interviews;
 	}
-
 	@JsonIgnore
 	public void setInterviews(Set<TfInterview> interviews) {
 		this.interviews = interviews;
@@ -112,7 +104,7 @@ public class TfEndClient implements java.io.Serializable {
 	public static long getSerialversionuid() {
 		return serialVersionUID;
 	}
-
+	
 	@Override
 	public String toString() {
 		return "TfEndClient [id=" + id + ", name=" + name + "]";
@@ -167,6 +159,4 @@ public class TfEndClient implements java.io.Serializable {
 		return true;
 	}
 	
-	
-
 }
