@@ -1,53 +1,70 @@
-//package com.revature.test.dao;
-//
-//import com.revature.dao.MarketingStatusDao;
-//import com.revature.model.MarketingStatusInfo;
-//import org.mockito.Mock;
-//import org.mockito.Mockito;
-//import org.mockito.MockitoAnnotations;
-//import org.testng.annotations.BeforeTest;
-//import org.testng.annotations.Test;
-//
-//import java.io.IOException;
-//import java.util.HashMap;
-//import java.util.Map;
-//
-//
-//public class MarketingStatusDAOTest {//extends BaseTest {
-//
-//    @Mock
-//    MarketingStatusDao msDao;
-//
-//
-//    Map<Integer, MarketingStatusInfo> mockMsMap;
-//
-//    @BeforeTest
-//    public void setupMocks() throws IOException {
-//        MockitoAnnotations.initMocks(this);
-//
-//        mockMsMap = new HashMap<>();
-//
-//        MarketingStatusInfo ms1 = new MarketingStatusInfo();
-//        ms1.setId(new Integer(1));
-//        ms1.setName("ms1");
-//
-//        MarketingStatusInfo ms2 = new MarketingStatusInfo();
-//        ms2.setId(new Integer(2));
-//        ms2.setName("ms2");
-//
-//        mockMsMap.put(ms1.getId(), ms1);
-//        mockMsMap.put(ms2.getId(), ms2);
-//        Mockito.when(msDao.getMarketingStatus())
-//                .thenReturn(mockMsMap);
-//    }
-//
-//    @Test
-//    public void testGetMarketingStatuses() throws Exception {
-//        
-//    }
-//
-//    @Test
-//    public void testGetAllMarketingStatuses() throws Exception {
-//        
-//    }
-//}
+package com.revature.test.dao;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import com.revature.daoimpl.MarketingStatusDaoImpl;
+import com.revature.entity.TfMarketingStatus;
+
+
+
+public class MarketingStatusDAOTest {
+
+  @Test
+  public void canGetMarketingStatusByValidID() {
+	  TfMarketingStatus statusObj = new MarketingStatusDaoImpl().getMarketingStatusById(1);
+	  Assert.assertEquals(statusObj.getName(), "MAPPED: TRAINING");
+  }
+  
+  @Test(expectedExceptions = NullPointerException.class)
+  public void invalidIDReturnsNullMarketingStatus() {  
+	  TfMarketingStatus statusObj = new MarketingStatusDaoImpl().getMarketingStatusById(1321312243);
+	  Assert.assertEquals(statusObj.getName(), null);
+  }
+  
+  @Test
+  public void allMarketingStatusDoesnotReturnEmptyList() {  
+	  List<TfMarketingStatus> list = new MarketingStatusDaoImpl().getAllMarketingStatuses();
+	  Assert.assertNotEquals(list.size(), 0);
+  }
+  
+  @Test
+  public void allMarketingStatusReturnsCorrectSizeList() {  
+	  List<TfMarketingStatus> list = new MarketingStatusDaoImpl().getAllMarketingStatuses();
+	  Assert.assertEquals(list.size(), 12);
+  }
+  
+  @Test
+  public void allMarketingListContainsAllStatuses() {  
+	  List<TfMarketingStatus> list = new MarketingStatusDaoImpl().getAllMarketingStatuses();
+	  List<String> statusesActual = new ArrayList<String>();
+	  for(TfMarketingStatus x : list) {
+		  statusesActual.add(x.getName());
+	  }
+	  String[] statuses = {"MAPPED: TRAINING",
+			  "MAPPED: RESERVED",
+			  "MAPPED: SELECTED",
+			  "MAPPED: CONFIRMED",
+			  "MAPPED: DEPLOYED",
+			  "UNMAPPED: TRAINING",
+			  "UNMAPPED: OPEN",
+			  "UNMAPPED: SELECTED",
+			  "UNMAPPED: CONFIRMED",
+			  "UNMAPPED: DEPLOYED",
+			  "DIRECTLY PLACED",
+			  "TERMINATED"};
+	  
+	  List<String> statusesExpected = new ArrayList<String>();
+	  
+	  for(String x : statuses) {
+		  statusesExpected.add(x);
+	  }
+
+	  Assert.assertTrue(statusesExpected.equals(statusesActual));
+  }
+  
+}
