@@ -2,13 +2,12 @@ package com.revature.test.junit.entity;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.assertNotEquals;
 
 import java.sql.Timestamp;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.junit.Test;
+import org.testng.annotations.Test;
 
 import com.revature.entity.TfAssociate;
 import com.revature.entity.TfBatch;
@@ -20,7 +19,7 @@ import com.revature.entity.TfPlacement;
 import com.revature.entity.TfUser;
 
 /**
- * Tests to test basic getter and setter functionality for TfAssociate
+ * Tests to test basic getter and GetSet functionality for TfAssociate
  * 
  * @author Jesse
  * @Since 6.18.06.11
@@ -42,78 +41,91 @@ public class TfAssociateTest {
 	TfAssociate tfassociate = new TfAssociate();
 
 	@Test
-	public void test2() {
+	public void testAssociateFirstNameGetSet() {
 		tfassociate.setFirstName("Bob");
 		assertTrue(tfassociate.getFirstName().equals("Bob"));
 		assertFalse(tfassociate.getFirstName().equals("bob"));
 	}
 
 	@Test
-	public void test3() {
+	public void testAssociateIDGetSet() {
 		tfassociate.setId(13);
 		assertTrue(tfassociate.getId() == 13);
 		assertFalse(tfassociate.getId() == 14);
 	}
 
 	@Test
-	public void test4() {
+	public void testAssociateLastNameGetSet() {
 		tfassociate.setLastName("Bobbert");
 		assertTrue(tfassociate.getLastName().equals("Bobbert"));
 		assertFalse(tfassociate.getLastName().equals("bobbert"));
 	}
 
 	@Test
-	public void test5() {
+	public void testAssociateBatchGetSet() {
 		tfassociate.setBatch(new TfBatch());
 		assertTrue(tfassociate.getBatch() instanceof TfBatch);
 	}
 
 	@Test
-	public void test6() {
+	public void testAssociateClientGetSet() {
 		tfassociate.setClient(new TfClient());
 		assertTrue(tfassociate.getClient() instanceof TfClient);
 	}
 
 	@Test
-	public void test7() {
+	public void testAssociateClientStartGetSet() {
 		tfassociate.setClientStartDate(new Timestamp(1000L));
 		assertTrue(tfassociate.getClientStartDate().getTime() == 1000L);
 		assertFalse(tfassociate.getClientStartDate().getTime() == 2000L);
 	}
 
 	@Test
-	public void test8() {
+	public void testAssociateEndClientGetSet() {
 		tfassociate.setEndClient(new TfEndClient());
 		assertTrue(tfassociate.getEndClient() instanceof TfEndClient);
 	}
 
 	@Test
-	public void test9() {
+	public void testAssociateInterviewGetSet() {
 		tfassociate.setInterview(new HashSet<TfInterview>());
 		assertTrue(tfassociate.getInterview() instanceof HashSet);
 	}
 
 	@Test
-	public void test10() {
+	public void testAssociateMarketStatusGetSet() {
 		tfassociate.setMarketingStatus(new TfMarketingStatus());
 		assertTrue(tfassociate.getMarketingStatus() instanceof TfMarketingStatus);
 	}
 
 	@Test
-	public void test11() {
+	public void testAssociatePlacementGetSet() {
 		tfassociate.setPlacement(new HashSet<TfPlacement>());
 		assertTrue(tfassociate.getPlacement() instanceof HashSet);
 	}
-
+	
 	@Test
-	public void test12() {
-		assertTrue(associate1.toString().equals(associate2.toString()));
-		assertFalse(associate1.equals(new TfAssociate()));
+	public void testAssociateFeedbackGetSet() {
+		String feedback1 = "For tfassociate";
+		String feedback2 = "For associate1";
+		tfassociate.setStagingFeedback(feedback1);
+		associate1.setStagingFeedback(feedback2);
+		assertTrue(tfassociate.getStagingFeedback().equals(feedback1));
+		assertTrue(!associate1.getStagingFeedback().equals(feedback1));
+	}
+	
+	//Have to hard code SerialID for now as it is private static,
+	//always verify this is using the correct SerialID as defined
+	//in the TfAssociate.java file
+	@Test
+	public void testAssociateSerialIDGet() {
+		long sid = -2324082555924677252L;
+		assertTrue(TfAssociate.getSerialversionuid() == sid);
 	}
 
 	@Test
-	public void test13() {
-		assertNotEquals(associate1.hashCode(), associate2.hashCode());
-		assertNotEquals(associate1.hashCode(), new TfAssociate().hashCode());
+	public void testAssociateEquivalence() {
+		assertTrue(associate1.toString().equals(associate2.toString()));
+		assertFalse(associate1.equals(new TfAssociate()));
 	}
 }
