@@ -27,34 +27,49 @@ public class TfCurriculumTest {
 	TfCurriculum tfcur = new TfCurriculum();
 
 	@Test
-	public void test1() {
+	public void testCurriculumBatchesGetSet() {
 		tfcur.setBatches(new HashSet<TfBatch>());
 		assertTrue(tfcur.getBatches() instanceof HashSet);
 	}
 
 	@Test
-	public void test2() {
+	public void testCurriculumIDGetSet() {
 		tfcur.setId(66);
 		assertTrue(tfcur.getId() == 66);
 		assertFalse(tfcur.getId() == 55);
 	}
 
 	@Test
-	public void test3() {
+	public void testCurriculumNameGetSet() {
 		tfcur.setName("Java");
 		assertTrue(tfcur.getName().equals("Java"));
 		assertFalse(tfcur.getName().equals("java"));
 	}
 
 	@Test
-	public void test4() {
+	public void testCurriculumEquivalence() {
 		assertTrue(cur1.equals(cur2));
 		assertFalse(cur1.equals(new TfCurriculum()));
 	}
 
 	@Test
-	public void test5() {
+	public void testCurriculumHashCode() {
 		assertEquals(cur1.hashCode(), cur2.hashCode());
 		assertNotEquals(cur1.hashCode(), new TfCurriculum().hashCode());
+	}
+	
+	//Have to hard code SerialID for now as it is private static,
+	//always verify this is using the correct SerialID as defined
+	//in the TfCurriculum.java file
+	@Test
+	public void testCurriculumSerialIDGet() {
+		long sid = 8213885869880424792L;
+		assertTrue(TfCurriculum.getSerialversionuid() == sid);
+	}
+	
+	@Test
+	public void testCurriculumToString() {
+		assertEquals(cur1.toString(), cur2.toString());
+		assertNotEquals(cur1.toString(), tfcur.toString());
 	}
 }
