@@ -26,53 +26,59 @@ public class TfTrainerTest {
 	TfTrainer trainer3 = new TfTrainer(10, "Jimbo", "Jimbo", new ArrayList<>(), new ArrayList<>());
 
 	@Test
-	public void test1() {
+	public void testTrainerCoTrainerGetSet() {
 		trainer.setCoTrainer(new ArrayList<>());
 		assertTrue(trainer.getCoTrainer() instanceof ArrayList);
 	}
 
 	@Test
-	public void test2() {
+	public void testTrainerFirstNameGetSet() {
 		trainer.setFirstName("Jim");
 		assertTrue(trainer.getFirstName().equals("Jim"));
 		assertFalse(trainer.getFirstName().equals("jim"));
 	}
 
 	@Test
-	public void test3() {
+	public void testTrainerIDGetSet() {
 		trainer.setId(7);
 		assertTrue(trainer.getId() == 7);
 		assertFalse(trainer.getId() == 6);
 	}
 
 	@Test
-	public void test4() {
+	public void testTrainerLastNameGetSet() {
 		trainer.setLastName("Carl");
 		assertTrue(trainer.getLastName().equals("Carl"));
 		assertFalse(trainer.getLastName().equals("carl"));
 	}
 
 	@Test
-	public void test5() {
+	public void testTrainerPrimaryGetSet() {
 		trainer.setPrimary(new ArrayList<>());
 		assertTrue(trainer.getPrimary() instanceof ArrayList);
 	}
 
 	@Test
-	public void test6() {
+	public void testTrainerTfUserGetSet() {
 		trainer.setTfUser(new TfUser());
 		assertTrue(trainer.getTfUser() instanceof TfUser);
 	}
 
 	@Test
-	public void test7() {
+	public void testTrainerEquivalence() {
 		assertTrue(trainer2.equals(trainer3));
 		assertFalse(trainer2.equals(trainer));
 	}
 
-	@Test(expectedExceptions = NullPointerException.class)
-	public void test8() {
+	@Test
+	public void testTrainerHashCode() {
 		assertEquals(trainer2.hashCode(), trainer3.hashCode());
 		assertNotEquals(trainer2.hashCode(), trainer.hashCode());
+	}
+	
+	@Test
+	public void testTrainerToString() {
+		assertEquals(trainer2.toString(), trainer3.toString());
+		assertNotEquals(trainer2.toString(), trainer.toString());
 	}
 }
