@@ -3,6 +3,7 @@ package com.revature.test.runner;
 import org.junit.runner.RunWith;
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
+import cucumber.api.testng.AbstractTestNGCucumberTests;
  
 /**
  * Test runner to run all of the features in the features attribute and to match them with the
@@ -12,10 +13,13 @@ import cucumber.api.junit.Cucumber;
  */
 @RunWith(Cucumber.class)
 @CucumberOptions(
-		features = {"src/test/resources/AdminFeatureFiles/AssociateView.feature","src/test/resources/AdminFeatureFiles/CreateUser.feature","src/test/resources/AdminFeatureFiles/HomeTab.feature","src/test/resources/AdminFeatureFiles/Login.feature","src/test/resources/AdminFeatureFiles/Predictions.feature","src/test/resources/AdminFeatureFiles/Register.feature"}
-		,glue = {"com.revature.test.utils","com.revature.test.admin.cukes"}
+		plugin = {"pretty"},
+		features = {"src/test/resources/AdminFeatureFiles"}
+		,glue = {"com.revature.test.utils","com.revature.test.admin.cuke"}
 		)
- 
-public class CucumberTestRunner {
+
+public class CucumberTestRunner extends AbstractTestNGCucumberTests {
  
 }
+// old code my add back later
+// glue = {"com.revature.test.utils","com.revature.test.admin.cuke"}
